Update ref namespace
During the review of the GitHub integration code (which is external to
this repository), WPT collaborators suggested that GitHub labels should
only be used to designate trust for submissions from non-collaborators.
Submissions from collaborators should be published by this system
without the automatic assignment of a GitHub label.

This change made the name `labeled-for-preview` inappropriate for the
corresponding namespace since the namespace will include patches that
have not been labeled via the GitHub UI.

Update the namespace to more accurately describe its contents.