Update the cert-renewers update policy & README
The cert-renewers had a different update policy. As a result, it needs
more steps to deploy it after the terraform apply. This change makes it
the same as the wpt live server.

Also update the README for process wise what to do. The terraform
changes should be approved before you do the deploy (which changes the
state file). Create a PR with the terraform change, get it approved.
Then do the deployment and create a separate PR with the state file
changes