[Part 3/3] Use latest OS image and use new metadata format (#60)
* [Part 3/3] Use latest OS image and use new metadata format

- Removes the hardcoded OS image so that terraform will build a template
  with the latest os image
- Previously, the metadata was hardcoded to ensure there were no diffs
  upon upgrading to the latest version of terraform. This change now
uses the new format for metadata for containers. (Which would have
caused a diff in part 1)

* Update the cert-renewers update policy & README

The cert-renewers had a different update policy. As a result, it needs
more steps to deploy it after the terraform apply. This change makes it
the same as the wpt live server.

Also update the README for process wise what to do. The terraform
changes should be approved before you do the deploy (which changes the
state file). Create a PR with the terraform change, get it approved.
Then do the deployment and create a separate PR with the state file
changes