Update the terraform state after building docker images
Following the instructions for deploying in the README.
terraform apply deployed the new docker images.

This was kept separate from the OS updates in #60. That updated the OS
version image. This updated the docker images for cert renewers and the
wpt.live server

All of these changes were generated automatically by terraform upon
apply