Migrate cert-renewal process to cloud run job
Previously, the cert-renewal process was a long standing instance that
ran the script once and slept. The problem arises that this sleep
eventually breaks and the instance never recovers. This migrates the
job to be a cloud run job that runs the script and that is it. The sleep
is now handled by a cron schedule. This ensures the instance is always
fresh.

Fixes #57
Fixes #77